move AArch64 specific code to llvm/include/llvm/Support/AArch64AttributeParser.h
